requirements.txt
argparse
python-nmap
tabulate
